Library

VectorEnumeration.base_qa — Function

base_qa(A::FreeAssAlgebra{FE}, R::Vector{AE}=AE[]) where {FE, AE <: FreeAssAlgElem{FE}}

Return a list of monomials of A forming a basis of the quotient algebra A/⟨ARA⟩ of the free associative algebra A and the 2-sided ideal generated by the subset R of A.

base_qa(A::FreeAssAlgebra{FE}, I::Oscar.FreeAssAlgIdeal{AE}) where {FE, AE <: FreeAssAlgElem{FE}}

Return a subset of A forming a k basis of the quotient algebra A/I of the free associative algebra A and the 2-sided ideal I.

VectorEnumeration.base_qm — Function

base_qm(A::FreeAssAlgebra{FE}, M::AbstractAlgebra.Module{AE}, R::Vector{AE}=AE[], W::Vector{ME}=ME[]) where {FE, AE <: FreeAssAlgElem{FE}, ME <: ModuleElem{AE}}

Return a list of monomials of the free module M = Aˢ forming a basis of the quotient module Pˢ/⟨πˢ(W)Pˢ⟩ for the quotient algebra P := A/⟨ARA⟩ of the free associative algebra A and the 2-sided ideal generated by the subset R of A, and the image π(W) of the subset W of M under the morphism πˢ : Aˢ → Pˢ.

Examples

julia> A, (a, b, c) = free_associative_algebra(GF(7), ["a", "b", "c"])
(Free associative algebra on 3 indeterminates over finite field F_7, AbstractAlgebra.Generic.FreeAssAlgElem{AbstractAlgebra.GFElem{Int64}}[a, b, c])

julia> M = free_module(A, 1)
Free module of rank 1 over free associative algebra on 3 indeterminates over finite field F_7

julia> R = [a^2 - 1, b^2 - 1, c^2 - 1, (a*b)^3 - 1, (a*c)^2 - 1, (b*c)^3 - 1]
6-element Vector{AbstractAlgebra.Generic.FreeAssAlgElem{AbstractAlgebra.GFElem{Int64}}}:
 a^2 + 6
 b^2 + 6
 c^2 + 6
 a*b*a*b*a*b + 6
 a*c*a*c + 6
 b*c*b*c*b*c + 6

julia> W = [M([a*b*c - 1])]
1-element Vector{AbstractAlgebra.Generic.FreeModuleElem{AbstractAlgebra.Generic.FreeAssAlgElem{AbstractAlgebra.GFElem{Int64}}}}:
 (a*b*c + 6)

julia> base_qm(A, M, R, W)
6-element Vector{AbstractAlgebra.Generic.FreeModuleElem{AbstractAlgebra.Generic.FreeAssAlgElem{AbstractAlgebra.GFElem{Int64}}}}:
 (1)
 (a)
 (a*b)
 (a*b*a)
 (b)
 (a*b*a*b)

base_qm(A::FreeAssAlgebra{FE}, s::Int=1, R::Vector{AE}=AE[], W::Vector{Vector{AE}}=Vector{AE}[]) where {FE, AE <: FreeAssAlgElem{FE}}

Return a list of mononomials of Aˢ forming a basis of the quotient module Pˢ/⟨πˢ(W)Pˢ⟩ for the quotient algebra P := A/⟨ARA⟩ of the free associative algebra A and the 2-sided ideal generated by the subset R of A, and the image π(W) of the subset W of the free module Aˢ of rank s over A under the morphism πˢ : Aˢ → Pˢ.

base_qm(A::FreeAssAlgebra{FE}, M::AbstractAlgebra.Module{AE}, I::Oscar.FreeAssAlgIdeal{AE}, W::Vector{ME}=ME[]) where {FE, AE <: FreeAssAlgElem{FE}, ME <: ModuleElem{AE}}

Return a subset of the free module M = Aˢ forming a basis of the quotient module Pˢ/⟨πˢ(W)Pˢ⟩ for the quotient algebra P := A/I of the free associative algebra A and the 2-sided ideal I, and the image π(W) of the subset W of M under the morphism πˢ : Aˢ → Pˢ.

base_qm(f::Field, FP::Oscar.FPGroup, U::Oscar.FPGroup=trivial_subgroup(FP))

Return a list of monomials forming a f - basis of f(FP/U)

Examples

julia> F = free_group(3)
<free group on the generators [ f1, f2, f3 ]>

julia> a, b, c = gens(F)
3-element Vector{FPGroupElem}:
 f1
 f2
 f3

julia> R = [a^2, b^2, c^2, (a*b)^3, (a*c)^2, (b*c)^3]
6-element Vector{FPGroupElem}:
 f1^2
 f2^2
 f3^2
 (f1*f2)^3
 (f1*f3)^2
 (f2*f3)^3

julia> N = normal_closure(F, sub(F, R)[1])[1]
Group(<free, no generators known>)

julia> FP = quo(FPGroup, F, N)[1]
<fp group of size 24 on the generators [ F1, F2, F3 ]>

julia> U = sub(FP, [FP[1]*FP[2]*FP[3]])[1]
Group([ F1*F2*F3 ])

julia> base_qm(QQ, FP, U)
6-element Vector{Vector{AbstractAlgebra.Generic.FreeAssAlgElem{QQFieldElem}}}:
 [1]
 [x1]
 [x1*x2]
 [x1*x2*x1]
 [x2]
 [x1*x2*x1*x2]

VectorEnumeration.dimension_qa — Function

dimension_qa(A::FreeAssAlgebra{FE}, R::Vector{AE}=AE[]) where {FE, AE <: FreeAssAlgElem{FE}}

Return the dimension of the quotient algebra A/⟨ARA⟩ of the free associative algebra A and the 2-sided ideal generated by the subset R of A.

dimension_qa(A::FreeAssAlgebra{FE}, I::Oscar.FreeAssAlgIdeal{AE}) where {FE, AE <: FreeAssAlgElem{FE}}

Return the dimension of the quotient algebra A/I of the free associative algebra A and the 2-sided ideal I.

VectorEnumeration.dimension_qm — Function

dimension_qm(A::FreeAssAlgebra{FE}, M::AbstractAlgebra.Module{AE}, R::Vector{AE}=AE[], W::Vector{ME}=ME[]) where {FE, AE <: FreeAssAlgElem{FE}, ME <: ModuleElem{AE}}

Return the dimension of the quotient module Pˢ/⟨πˢ(W)Pˢ⟩ for the quotient algebra P := A/⟨ARA⟩ of the free associative algebra A and the 2-sided ideal generated by the subset R of A, and the image π(W) of the subset W of the free module M = Aˢ under the morphism πˢ : Aˢ → Pˢ.

Examples

julia> A, (a, b, c) = free_associative_algebra(GF(7), ["a", "b", "c"])
(Free associative algebra on 3 indeterminates over finite field F_7, AbstractAlgebra.Generic.FreeAssAlgElem{AbstractAlgebra.GFElem{Int64}}[a, b, c])

julia> M = free_module(A, 1)
Free module of rank 1 over free associative algebra on 3 indeterminates over finite field F_7

julia> R = [a^2 - 1, b^2 - 1, c^2 - 1, (a*b)^3 - 1, (a*c)^2 - 1, (b*c)^3 - 1]
6-element Vector{AbstractAlgebra.Generic.FreeAssAlgElem{AbstractAlgebra.GFElem{Int64}}}:
 a^2 + 6
 b^2 + 6
 c^2 + 6
 a*b*a*b*a*b + 6
 a*c*a*c + 6
 b*c*b*c*b*c + 6

julia> W = [M([a*b*c - 1])]
1-element Vector{AbstractAlgebra.Generic.FreeModuleElem{AbstractAlgebra.Generic.FreeAssAlgElem{AbstractAlgebra.GFElem{Int64}}}}:
 (a*b*c + 6)

julia> dimension_qm(A, M, R, W)
6

dimension_qm(A::FreeAssAlgebra{FE}, s::Int=1, R::Vector{AE}=AE[], W::Vector{Vector{AE}}=Vector{AE}[]) where {FE, AE <: FreeAssAlgElem{FE}}

Return the dimension of the quotient module Pˢ/⟨πˢ(W)Pˢ⟩ for the quotient algebra P := A/⟨ARA⟩ of the free associative algebra A and the 2-sided ideal generated by the subset R of A, and the image π(W) of the subset W of the free module Aˢ of rank s over A under the morphism πˢ : Aˢ → Pˢ.

dimension_qm(A::FreeAssAlgebra{FE}, M::AbstractAlgebra.Module{AE}, I::Oscar.FreeAssAlgIdeal{AE}, W::Vector{ME}=ME[]) where {FE, AE <: FreeAssAlgElem{FE}, ME <: ModuleElem{AE}}

Return the dimension of the quotient module Pˢ/⟨πˢ(W)Pˢ⟩ for the quotient algebra P := A/I of the free associative algebra A and the 2-sided ideal I, and the image π(W) of the subset W of the free module M = Aˢ under the morphism πˢ : Aˢ → Pˢ.

dimension_qm(f::Field, FP::Oscar.FPGroup, U::Oscar.FPGroup=trivial_subgroup(FP))

Return the f - dimension of f(FP/U)

Examples

julia> F = free_group(3)
<free group on the generators [ f1, f2, f3 ]>

julia> a, b, c = gens(F)
3-element Vector{FPGroupElem}:
 f1
 f2
 f3

julia> R = [a^2, b^2, c^2, (a*b)^3, (a*c)^2, (b*c)^3]
6-element Vector{FPGroupElem}:
 f1^2
 f2^2
 f3^2
 (f1*f2)^3
 (f1*f3)^2
 (f2*f3)^3

julia> N = normal_closure(F, sub(F, R)[1])[1]
Group(<free, no generators known>)

julia> FP= quo(FPGroup, F, N)[1]
<fp group of size 24 on the generators [ F1, F2, F3 ]>

julia> U = sub(FP, [FP[1]*FP[2]*FP[3]])[1]
Group([ F1*F2*F3 ])

julia> dimension_qm(QQ, FP, U)
6

VectorEnumeration.matrices_qa — Function

matrices_qa(T::Type, A::FreeAssAlgebra{FE}, R::Vector{AE}=AE[]) where {FE, AE <: FreeAssAlgElem{FE}}

Return matrix representations as matrices of type T for the generators of the free associative algebra A for right multiplication in the quotient algebra A/⟨ARA⟩ of A and the 2-sided ideal generated by the subset R of A.

matrices_qa(T::Type, A::FreeAssAlgebra{FE}, I::Oscar.FreeAssAlgIdeal{AE}) where {FE, AE <: FreeAssAlgElem{FE}}

Return matrix representations as matrices of type T for the generators of the free associative algebra A for right multiplication in the quotient algebra A/I of A and the 2-sided ideal I.

VectorEnumeration.matrices_qm — Function

matrices_qm(T::Type, A::FreeAssAlgebra{FE}, M::AbstractAlgebra.Module{AE}, R::Vector{AE}=AE[], W::Vector{ME}=ME[]) where {FE, AE <: FreeAssAlgElem{FE}, ME <: ModuleElem{AE}}

Return matrix representations as matrices of type T for the generators of the free associative algebra A for right multiplication in the quotient module Pˢ/⟨πˢ(W)Pˢ⟩ for the quotient algebra P := A/⟨ARA⟩ of A and the 2-sided ideal generated by the subset R of A, and the image π(W) of the subset W of the free module M = Aˢ under the morphism πˢ : Aˢ → Pˢ.

Examples

julia> A, (a, b, c) = free_associative_algebra(GF(7), ["a", "b", "c"])
(Free associative algebra on 3 indeterminates over finite field F_7, AbstractAlgebra.Generic.FreeAssAlgElem{AbstractAlgebra.GFElem{Int64}}[a, b, c])

julia> M = free_module(A, 1)
Free module of rank 1 over free associative algebra on 3 indeterminates over finite field F_7

julia> R = [a^2 - 1, b^2 - 1, c^2 - 1, (a*b)^3 - 1, (a*c)^2 - 1, (b*c)^3 - 1]
6-element Vector{AbstractAlgebra.Generic.FreeAssAlgElem{AbstractAlgebra.GFElem{Int64}}}:
 a^2 + 6
 b^2 + 6
 c^2 + 6
 a*b*a*b*a*b + 6
 a*c*a*c + 6
 b*c*b*c*b*c + 6

julia> W = [M([a*b*c - 1])]
1-element Vector{AbstractAlgebra.Generic.FreeModuleElem{AbstractAlgebra.Generic.FreeAssAlgElem{AbstractAlgebra.GFElem{Int64}}}}:
 (a*b*c + 6)

julia> Xᵃ, Xᵇ, Xᶜ = matrices_qm(Matrix, A, M, R, W)
3-element Vector{Matrix{AbstractAlgebra.GFElem{Int64}}}:
 [0 1 … 0 0; 1 0 … 0 0; … ; 0 0 … 0 1; 0 0 … 1 0]
 [0 0 … 1 0; 0 0 … 0 0; … ; 1 0 … 0 0; 0 0 … 0 0]
 [0 0 … 0 0; 0 0 … 0 0; … ; 0 0 … 0 1; 0 0 … 1 0]

julia> Xᵃ
6×6 Matrix{AbstractAlgebra.GFElem{Int64}}:
 0  1  0  0  0  0
 1  0  0  0  0  0
 0  0  0  1  0  0
 0  0  1  0  0  0
 0  0  0  0  0  1
 0  0  0  0  1  0

julia> Xᵇ
6×6 Matrix{AbstractAlgebra.GFElem{Int64}}:
 0  0  0  0  1  0
 0  0  1  0  0  0
 0  1  0  0  0  0
 0  0  0  0  0  1
 1  0  0  0  0  0
 0  0  0  1  0  0

julia> Xᶜ
6×6 Matrix{AbstractAlgebra.GFElem{Int64}}:
 0  0  1  0  0  0
 0  0  0  1  0  0
 1  0  0  0  0  0
 0  1  0  0  0  0
 0  0  0  0  0  1
 0  0  0  0  1  0

matrices_qm(T::Type, A::FreeAssAlgebra{FE}, s::Int=1, R::Vector{AE}=AE[], W::Vector{Vector{AE}}=Vector{AE}[]) where {FE, AE <: FreeAssAlgElem{FE}}

Return matrix representations as matrices of type T for the generators of the free associative algebra A for right multiplication in the quotient module Pˢ/⟨πˢ(W)Pˢ⟩ for the quotient algebra P := A/⟨ARA⟩ of A and the 2-sided ideal generated by the subset R of A, and the image π(W) of the subset W of the free module Aˢ of rank s over A under the morphism πˢ : Aˢ → Pˢ.

matrices_qm(T::Type, A::FreeAssAlgebra{FE}, M::AbstractAlgebra.Module{AE}, I::Oscar.FreeAssAlgIdeal{AE}, W::Vector{ME}=ME[]) where {FE, AE <: FreeAssAlgElem{FE}, ME <: ModuleElem{AE}}

Return matrix representations as matrices of type T for the generators of the free associative algebra A for right multiplication in the quotient module Pˢ/⟨πˢ(W)Pˢ⟩ for the quotient algebra P := A/I of A and the 2-sided ideal I, and the image π(W) of the subset W of the free module M = Aˢ under the morphism πˢ : Aˢ → Pˢ.

matrices_qm(T::Type, f::Field, FP::Oscar.FPGroup, U::Oscar.FPGroup=trivial_subgroup(FP))

Return a f - matrix - representation of f(FP/U) in matrices of type T

Examples

julia> F = free_group(3)
<free group on the generators [ f1, f2, f3 ]>

julia> a, b, c = gens(F)
3-element Vector{FPGroupElem}:
 f1
 f2
 f3

julia> R = [a^2, b^2, c^2, (a*b)^3, (a*c)^2, (b*c)^3]
6-element Vector{FPGroupElem}:
 f1^2
 f2^2
 f3^2
 (f1*f2)^3
 (f1*f3)^2
 (f2*f3)^3

julia> N = normal_closure(F, sub(F, R)[1])[1]
Group(<free, no generators known>)

julia> FP= quo(FPGroup, F, N)[1]
<fp group of size 24 on the generators [ F1, F2, F3 ]>

julia> U = sub(FP, [FP[1]*FP[2]*FP[3]])[1]
Group([ F1*F2*F3 ])

julia> Xᵃ, Xᵇ, Xᶜ = matrices_qm(Matrix, QQ, FP, U)
3-element Vector{Matrix{QQFieldElem}}:
 [0 1 … 0 0; 1 0 … 0 0; … ; 0 0 … 0 1; 0 0 … 1 0]
 [0 0 … 1 0; 0 0 … 0 0; … ; 1 0 … 0 0; 0 0 … 0 0]
 [0 0 … 0 0; 0 0 … 0 0; … ; 0 0 … 0 1; 0 0 … 1 0]

julia> Xᵃ
6×6 Matrix{QQFieldElem}:
 0  1  0  0  0  0
 1  0  0  0  0  0
 0  0  0  1  0  0
 0  0  1  0  0  0
 0  0  0  0  0  1
 0  0  0  0  1  0

julia> Xᵇ
6×6 Matrix{QQFieldElem}:
 0  0  0  0  1  0
 0  0  1  0  0  0
 0  1  0  0  0  0
 0  0  0  0  0  1
 1  0  0  0  0  0
 0  0  0  1  0  0

julia> Xᶜ
6×6 Matrix{QQFieldElem}:
 0  0  1  0  0  0
 0  0  0  1  0  0
 1  0  0  0  0  0
 0  1  0  0  0  0
 0  0  0  0  0  1
 0  0  0  0  1  0